KL-147/rename: rename SubCategoryContainer to SubCategoryArray
SubCategoryContainer 컴포넌트의 이름이 변경됨에 따라 파일명도 동일하게 변경하였습니다.
- SubCategoryArray의 import 경로를 수정하였습니다.
